requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
sqlalchemy
pydantic
python-dotenv
